requirements.txt
Django==1.10.4
django-allauth==0.29.0
django-bootstrap-form==3.2.1
django-filter==1.0.1
django-rest-api==0.1.5
djangorestframework==3.5.3
drfdocs==0.0.11
gunicorn==19.6.0
oauthlib==2.0.1
python-openid==2.2.5
requests==2.12.4
requests-oauthlib==0.7.0
whitenoise==3.2.2
